% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplotbl.R
\name{coefplotbl}
\alias{coefplotbl}
\title{Coefficient plot with table}
\usage{
coefplotbl(
  x,
  intercept = FALSE,
  add_tbl = TRUE,
  statistics = "ci",
  alpha = 0.05,
  sig = TRUE,
  digits = 3,
  coef_rename = NULL,
  coef_omit = NULL,
  highlight = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  size = 1,
  linewidth = 0.75,
  fontsize = 12,
  colors = c(sig = "black", insig = "gray70", highlight = "red")
)
}
\arguments{
\item{x}{a \code{lm} or \code{glm} object.}

\item{intercept}{a logical. Default is \code{FALSE}.}

\item{add_tbl}{a logical. Default is \code{TRUE}.}

\item{statistics}{\code{"ci"} (confidential interval; default), \code{"se"} (standard error), \code{"t"} (t-value), or \code{"p"} (p-value).}

\item{alpha}{Default is \code{0.05}.}

\item{sig}{If \code{TRUE} (default), the point-ranges which are statistically significant are highlighted.}

\item{digits}{Default is \code{3}.}

\item{coef_rename}{A named character vector.}

\item{coef_omit}{A character vector.}

\item{highlight}{A character vector.}

\item{xlab}{a character.}

\item{ylab}{a character.}

\item{title}{a character.}

\item{size}{a numeric value. Default is \code{1}.}

\item{linewidth}{A numeric value. Default is \code{0.75}.}

\item{fontsize}{A numeric value. Default is \code{12}.}

\item{colors}{A named character vector. The names must be \code{sig}, \code{insig}, and \code{highlight}.}
}
\value{
a \code{ggplot} object
}
\description{
Coefficient plot with table
}
\examples{
# Example 1
fit <- lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width + Species, data = iris)
coefplotbl(fit)

# Example 2: Display only plot & 90\\% CI
coefplotbl(fit, add_tbl = FALSE, alpha = 0.1)

# Example 3: Recoding covariate names
coefplotbl(fit, coef_rename = c("Sepal.Width"  = "Sepal Width",
                                "Petal.Length" = "Petal Length",
                                "Petal.Width"  = "Petal Width"))

# Example 4: Highlight specific covariates
coefplotbl(fit, coef_rename = c("Sepal.Width"  = "Sepal Width",
                                "Petal.Length" = "Petal Length",
                                "Petal.Width"  = "Petal Width"),
           highlight = c("Petal.Length", "Petal.Width"))

# Example 5: Omit specific covariates
coefplotbl(fit, coef_omit = c("Sepal.Width"))

# Example 6: Display p-value instead of confidential interval (4-digit)
coefplotbl(fit, statistics = "p", digits = 4)

}
