% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplotbl.R
\name{coefplotbl}
\alias{coefplotbl}
\title{Coefficient plot with table}
\usage{
coefplotbl(
  x,
  intercept = FALSE,
  add_tbl = TRUE,
  statistics = "ci",
  alpha = 0.05,
  sig = TRUE,
  digits = 3,
  coef_rename = NULL,
  coef_omit = NULL,
  coef_order = NULL,
  highlight = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  size = 1,
  linewidth = 0.75,
  fontsize = 12,
  colors = c(sig = "black", insig = "gray70", highlight = "red"),
  ...
)
}
\arguments{
\item{x}{a \code{lm} or \code{glm} object. Objects available for the \code{tidy()} in the \{broom\} package.}

\item{intercept}{a logical. If \code{TRUE}, the intercept is displayed.Default is \code{FALSE}.}

\item{add_tbl}{a logical. If \code{TRUE} (default), a regression table appears on the right side of the plot.}

\item{statistics}{\code{"ci"} (confidential interval; default), \code{"se"} (standard error), \code{"t"} (t-value), or \code{"p"} (p-value).}

\item{alpha}{a numeric. a significant level. Default is \code{0.05}.}

\item{sig}{a logical. If \code{TRUE} (default), the point-ranges which are statistically significant are highlighted.}

\item{digits}{a numeric. number of digits. Default is \code{3}.}

\item{coef_rename}{A named character vector. For example, \code{c("Old1" = "New1", "Old2" = "New2")}}

\item{coef_omit}{A character vector. Covariate names to be omitted. The covariate names must be the name before it was changed by \code{coef_rename}.}

\item{coef_order}{A character vector. Order of covariates. The covariate names must be the name before it was changed by \code{coef_rename} and must include \code{"(Intercept)"}.}

\item{highlight}{A character vector. Covariate names to be highlighted. The covariate names must be the name before it was changed by \code{coef_rename}.}

\item{xlab}{a character.}

\item{ylab}{a character.}

\item{title}{a character.}

\item{size}{a numeric value. Point size. Default is \code{1}.}

\item{linewidth}{A numeric value. Line width. Default is \code{0.75}.}

\item{fontsize}{A numeric value. Default is \code{12}.}

\item{colors}{A named character vector. The names must be \code{sig}, \code{insig}, and \code{highlight}.}

\item{...}{ignored.}
}
\value{
a \code{ggplot} object
}
\description{
Coefficient plot with table
}
\examples{
# Example 1
fit <- lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width + Species, data = iris)
coefplotbl(fit)

# Example 2: Display only plot & 99\% CI
coefplotbl(fit, add_tbl = FALSE, alpha = 0.01)

# Example 3: Recoding covariate names
coefplotbl(fit, coef_rename = c("Sepal.Width"  = "Sepal Width",
                                "Petal.Length" = "Petal Length",
                                "Petal.Width"  = "Petal Width"))

# Example 4: Highlight specific covariates
coefplotbl(fit, coef_rename = c("Sepal.Width"  = "Sepal Width",
                                "Petal.Length" = "Petal Length",
                                "Petal.Width"  = "Petal Width"),
           highlight = c("Petal.Length", "Petal.Width"))

# Example 5: Omit specific covariates
coefplotbl(fit, coef_omit = c("Sepal.Width"))

# Example 6: Display p-value instead of confidential interval (4-digit)
coefplotbl(fit, statistics = "p", digits = 4)

# Example 7: Labels
coefplotbl(fit, alpha = 0.01, title = "Title",
           xlab = "x-axis label", ylab = "y-axis label")

# Example 8: Colors
coefplotbl(fit,
           coef_rename = c("Sepal.Width"  = "Sepal Width",
                           "Petal.Length" = "Petal Length",
                           "Petal.Width"  = "Petal Width",
                           "Speciesversicolor" = "Versicolor",
                           "Speciesvirginica"  = "Virginica"),
           highlight   = c("Speciesversicolor", "Speciesvirginica"),
           alpha       = 0.01,
           colors      = c(sig = "black", insig = "gray90", highlight = "royalblue"))

# Example 9: Show intercept
coefplotbl(fit, intercept = TRUE)

# Example 10: Re-order covariates
coefplotbl(fit, coef_order = c("Speciesversicolor", "Speciesvirginica",
                               "Sepal.Width", "Petal.Width", "Petal.Length",
                               "(Intercept)"))

}
